CP/python-scripts/requirements.txt
PyJWT==2.8.0
qrcode[pil]==7.4.2
Pillow==10.0.1
cryptography==41.0.7